more silence warnings when using -Wconversion
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@2588 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/lib/dim.c
@@ -87,7 +87,7 @@ ncmpii_new_NC_dim(NC_dimarray  *ncap,
         /* allocate or expand the space for nameT[key].list */
         if (nameT[key].num % NC_NAME_TABLE_CHUNK == 0)
             nameT[key].list = (int*) NCI_Realloc(nameT[key].list,
-                              (nameT[key].num+NC_NAME_TABLE_CHUNK) * sizeof(int));
+                              (size_t)(nameT[key].num+NC_NAME_TABLE_CHUNK) * sizeof(int));
 
         /* add the new variable ID to the name lookup table
          * the new varid will be ncap->ndefined
@@ -284,9 +284,9 @@ ncmpii_dup_NC_dimarray(NC_dimarray *ncap, const NC_dimarray *ref)
         ncap->nameT[i].num = ref->nameT[i].num;
         ncap->nameT[i].list = NULL;
         if (ncap->nameT[i].num > 0) {
-            ncap->nameT[i].list = NCI_Malloc(ncap->nameT[i].num * sizeof(int));
+            ncap->nameT[i].list = NCI_Malloc((size_t)ncap->nameT[i].num * sizeof(int));
             memcpy(ncap->nameT[i].list, ref->nameT[i].list,
-                   ncap->nameT[i].num * sizeof(int));
+                   (size_t)ncap->nameT[i].num * sizeof(int));
         }
     }
