+++ src/lib/ncmpidtype.c
@@ -356,9 +356,9 @@ int ncmpii_dtype_decode(MPI_Datatype dtype,
     return NC_NOERR;
   }
 
-  array_of_ints = (int *) NCI_Malloc(num_ints * SIZEOF_INT);
-  array_of_adds = (MPI_Aint *) NCI_Malloc(num_adds * SIZEOF_MPI_AINT);
-  array_of_dtypes = (MPI_Datatype *) NCI_Malloc(num_dtypes * sizeof(MPI_Datatype));
+  array_of_ints = (int *) NCI_Malloc((size_t)num_ints * SIZEOF_INT);
+  array_of_adds = (MPI_Aint *) NCI_Malloc((size_t)num_adds * SIZEOF_MPI_AINT);
+  array_of_dtypes = (MPI_Datatype *) NCI_Malloc((size_t)num_dtypes * sizeof(MPI_Datatype));
 
   MPI_Type_get_contents(dtype, num_ints, num_adds, num_dtypes,
                         array_of_ints, array_of_adds, array_of_dtypes);
